refactored kitchen
